Show all analysis on completion of first experiment
